Prevent bleeding methods into main scope
namespaces in Rake are purely for organizing the tasks, they do not
open a new lexical scope. Therefore, all methods defined here are
bound to Kernel (the implicit receiver). This makes them available
to ALL rake-task and the application if the rake-tasks are being
loaded. It is better to wrap them in a module and call them there.
To avoid further shenanigans, 'module_function' allows calling them
as if they were class-methods.